(codethorn) changed TransitionGraph hash function to use size_t only
CODETHORN-70

+++ projects/CodeThorn/src/TransitionGraph.C
@@ -30,10 +30,10 @@ size_t Transition::memorySize() const {
 TransitionHashFun::TransitionHashFun() {
 }
 
-long TransitionHashFun::operator()(Transition* s) const {
-  unsigned int hash=1;
-  hash=((((long)s->source)+1)<<8)+(long)s->target*(long)s->edge.hash();
-  return long(hash);
+size_t TransitionHashFun::operator()(Transition* s) const {
+  size_t hash=1;
+  hash=((((size_t)s->source)+1)<<8)+(size_t)s->target*(size_t)s->edge.hash();
+  return hash;
 }
 
 TransitionEqualToPred::TransitionEqualToPred() {}
